requirement.txt
google-genai==0.3.0
pymongo==4.15.5
chromadb==1.3.5
requests==2.32.3
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1